+++ example/tests/test_notify.py
@@ -23,6 +23,7 @@ class NotifyTestCase(TestCase):
             title='{{greeting}} {{user}}!',
             text='Welcome to PyNotify!',
             trigger_action='http://localhost/',
+            extra_fields={'abc': '123'},
             related_objects={'user': self.user},
             extra_data={'greeting': 'Hello'},
             dispatcher_classes=(MockDispatcher,)
@@ -33,6 +34,7 @@ class NotifyTestCase(TestCase):
         self.assertEqual(notification.title, 'Hello John!')
         self.assertEqual(notification.text, 'Welcome to PyNotify!')
         self.assertEqual(notification.trigger_action, 'http://localhost/')
+        self.assertEqual(notification.extra_fields['abc'], '123')
         MockDispatcher.dispatch.assert_called_once_with(notification)
 
     def test_notify_should_create_notification_from_existing_template(self):